use type long to store time instead of double

+++ src/shooter.h
@@ -27,7 +27,7 @@ class C_Shooter : public C_GameUnits
 	protected:
 	C_Weapon *m_weapon;
 	SDL_Texture* m_image;
-	double m_lastShootTime;
+	long m_lastShootTime;
 };
 
 #endif
